requirements.txt
livekit-agents>=1.1.0
livekit-plugins>=1.1.0
openai>=1.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
tzdata>=2023.3
speexdsp-python
av